divide diagram context menu, add stub context menu for scheme screen
