+++ src/renderer/src/components/DiagramContextMenu/SchemeScreenContextMenu.tsx
@@ -1,0 +1,8 @@
+/*
+  Контекстное меню при работе со схемотехническим экраном.
+*/
+export const SchemeScreenContextMenu: React.FC = () => {
+  // TODO(L140-beep): Контекстное меню для схемотехнического экрана
+  console.log('heree');
+  return <></>;
+};
